requirements.txt
bigvgan
swanlab
vector_quantize_pytorch
timm
torchaudio
transformers